requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==2.2.13
django-cors-headers==3.5.0
djangorestframework==3.11.0
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
pycodestyle==2.6.0
pylint==2.6.0
pytz==2020.1
razorpay==1.2.0
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
